login admin and dealer both allowed add icons

+++ Natural Veneer Phase 1 21 July/VeeNilam/CART/CartCell/CartViewCell.m
@@ -31,6 +31,6 @@
 }
 -(IBAction)btnRemoveCartTapped:(UIButton *)sender
 {
-    [self.delegate btnRemoveFromCartTapped:(sender.tag - 1)];
+    [self.delegate btnRemoveFromCartTapped:(int)(sender.tag - 1)];
 }
 @end
